% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{dtsal}
\alias{dtsal}
\title{Calculate the probability density
Input: vector of data values, distributional parameters, left-censoring
       threshold, log flag
Output: vector of (log) densities}
\usage{
dtsal(
  x,
  shape = 1,
  scale = 1,
  q = tsal.q.from.shape(shape),
  kappa = tsal.kappa.from.ss(shape, scale),
  xmin = 0,
  log = FALSE
)
}
\description{
Calculate the probability density
Input: vector of data values, distributional parameters, left-censoring
       threshold, log flag
Output: vector of (log) densities
}
