% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gThetaFunction.R
\name{g.theta.inv}
\alias{g.theta.inv}
\title{\eqn{g(\theta)} function inverse}
\usage{
g.theta.inv(value, lower = 0, upper = 10000, v, u = -v)
}
\arguments{
\item{v}{Tuning parameter}

\item{u}{Tuning parameter}

\item{theta}{Parameter \eqn{\theta} of qExponential function}
}
\description{
Calculate the inverse of the function \eqn{g(\theta)}.
}
