% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gThetaFunction.R
\name{g.theta}
\alias{g.theta}
\title{\eqn{g(\theta)} function}
\usage{
g.theta(theta, v, u = -v)
}
\arguments{
\item{theta}{Parameter \eqn{\theta} of qExponential function}

\item{v}{Tuning parameter}

\item{u}{Tuning parameter}
}
\description{
Function to calculate \eqn{g(\theta)}.
}
