% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{K.mu.theta}
\alias{K.mu.theta}
\title{Expected Fisher Information}
\usage{
K.mu.theta(mu, theta)
}
\arguments{
\item{mu}{True value of mu}

\item{theta}{True value of theta}
}
\description{
Expected Fisher Information
}
