% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{KMuTheta}
\alias{KMuTheta}
\title{Expected Fisher Information}
\usage{
KMuTheta(mu, theta)
}
\arguments{
\item{mu}{True value of mu}

\item{theta}{True value of theta}
}
\description{
Expected Fisher Information
}
