% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gThetaFunction.R
\name{limits.g}
\alias{limits.g}
\title{Limits of \eqn{g(\theta)} function}
\usage{
limits.g(v, u)
}
\arguments{
\item{v}{Tuning parameter}

\item{u}{Tuning parameter}
}
\value{
A list of two elements: limit of \eqn{g(\theta)} when \eqn{\theta} goes to 0 and limit of \eqn{g(\theta)} when \eqn{\infty}
}
\description{
Limits of \eqn{g(\theta)} function
}
