% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{log.qexp}
\alias{log.qexp}
\title{Log of likelihood qexp}
\usage{
\method{log}{qexp}(param, n)
}
\arguments{
\item{param}{vector of paramters \eqn{(\mu,\theta)}}

\item{n}{sample size}
}
\description{
Log of likelihood qexp
}
