% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{logQexp}
\alias{logQexp}
\title{Log of likelihood qexp}
\usage{
logQexp(param, x, n)
}
\arguments{
\item{param}{vector of paramters \eqn{(\mu,\theta)}}

\item{n}{sample size}
}
\description{
Log of likelihood qexp
}
