% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfmmQuantities.R
\name{mfmm.theo}
\alias{mfmm.theo}
\title{Theoretical quantities MFMM}
\usage{
mfmm.theo(mu, theta, v, u, prob = 0.97)
}
\arguments{
\item{mu}{true value of mu}

\item{theta}{true value of theta}

\item{v}{Tuning parameter}

\item{u}{Tuning parameter}
}
\description{
Theoretical quantities for the mfmm
}
