% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfmmQuantities.R
\name{mfmmTheo}
\alias{mfmmTheo}
\title{Theoretical quantities MFMM}
\usage{
mfmmTheo(
  mu,
  theta,
  v,
  u,
  prob = 0.97,
  n = 100,
  d = 1e-10,
  lower = 0,
  upper = 10000
)
}
\arguments{
\item{mu}{true value of mu}

\item{theta}{true value of theta}

\item{v}{Tuning parameter}

\item{u}{Tuning parameter}
}
\description{
Theoretical quantities for the mfmm
}
