% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{mle.samples}
\alias{mle.samples}
\title{Sample quantities MLE}
\usage{
mle.samples(samples, mu, theta)
}
\arguments{
\item{samples}{matrix of samples Nxn}

\item{mu}{true value of mu}

\item{theta}{true value of theta}
}
\description{
Theoretical quantities for the mle
}
