% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{mleSamples}
\alias{mleSamples}
\title{Sample quantities MLE}
\usage{
mleSamples(samples, mu, theta)
}
\arguments{
\item{samples}{matrix of samples Nxn}

\item{mu}{true value of mu}

\item{theta}{true value of theta}
}
\description{
Theoretical quantities for the mle
}
