% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCompDens.mfmm}
\alias{plotCompDens.mfmm}
\title{Plot densities
Comparison of densities of mfmm estimators and Standart Normal}
\usage{
plotCompDens.mfmm(data, n, theta, parameter = "theta", standard = TRUE)
}
\arguments{
\item{data}{data frame}

\item{n}{vector of samples sizes}

\item{theta}{vector of values of theta (as character)}

\item{parameter}{if is "theta" or "q"}

\item{standard}{if TRUE plot the standard stimates}
}
\description{
Plot densities
Comparison of densities of mfmm estimators and Standart Normal
}
