% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{points.mfmm}
\alias{points.mfmm}
\title{points mfmm
Plot points}
\usage{
\method{points}{mfmm}(mfmm.est, v, u, xlimit = 100, colour = "black", size = 0.25)
}
\arguments{
\item{mfmm.est}{object obtained from mfmm.samples function}

\item{v}{tuning parameter}

\item{u}{tuning parameter}
}
\description{
points mfmm
Plot points
}
