% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{ptsal}
\alias{ptsal}
\title{Calculate the cumulative distribution function
Input: vector of data values, distributional parameters, left-censoring
       threshold, usual flags
Output: vector of (log) probabilities}
\usage{
ptsal(
  x,
  shape = 1,
  scale = 1,
  q = tsal.q.from.shape(shape),
  kappa = tsal.kappa.from.ss(shape, scale),
  xmin = 0,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\description{
Calculate the cumulative distribution function
Input: vector of data values, distributional parameters, left-censoring
       threshold, usual flags
Output: vector of (log) probabilities
}
