% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExpSamples.R
\name{qexp.samples}
\alias{qexp.samples}
\title{Samples generator}
\usage{
qexp.samples(N, n, theta, mu)
}
\arguments{
\item{N}{Number of samples}

\item{n}{Sample size of each sample}

\item{theta}{True value of theta}

\item{mu}{True value of mu}
}
\value{
A matrix of order N x n, where each row is a sample of size n.
}
\description{
Generates samples in the parameterization in terms of \eqn{\mu} and \eqn{\theta} (Patriota, 2012)
}
