% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{qtsal}
\alias{qtsal}
\title{Calculate quantiles
Input: vector of p-values, distributional parameters, left-censoring
       threshold, usual flags
Output: vector of quantile locations}
\usage{
qtsal(
  p,
  shape = 1,
  scale = 1,
  q = tsal.q.from.shape(shape),
  kappa = tsal.kappa.from.ss(shape, scale),
  xmin = 0,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\description{
Calculate quantiles
Input: vector of p-values, distributional parameters, left-censoring
       threshold, usual flags
Output: vector of quantile locations
}
