% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{reject.mfmm}
\alias{reject.mfmm}
\title{Reject mfmm}
\usage{
reject.mfmm(mu, theta, nrow = 1)
}
\arguments{
\item{mu}{paramter mu}

\item{theta}{vector of values of theta}

\item{nrow}{number of rows of facet grid}
}
\description{
Plot asymptotic approximation for different values of n
}
