% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{rtsal}
\alias{rtsal}
\title{Generate random variates
Input: integer length, distributional parameters, left-censoring threshold
Output: vector of reals}
\usage{
rtsal(
  n,
  shape = 1,
  scale = 1,
  q = tsal.q.from.shape(shape),
  kappa = tsal.kappa.from.ss(shape, scale),
  xmin = 0
)
}
\description{
Generate random variates
Input: integer length, distributional parameters, left-censoring threshold
Output: vector of reals
}
