% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{tsal.kappa.from.ss}
\alias{tsal.kappa.from.ss}
\title{Calculate kappa parameter from shape and scale parameters
Input: two real values
Output: a real value}
\usage{
tsal.kappa.from.ss(shape, scale)
}
\description{
Calculate kappa parameter from shape and scale parameters
Input: two real values
Output: a real value
}
