% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{tsal.q.from.shape}
\alias{tsal.q.from.shape}
\title{Calculate q parameter from shape parameter
Input: a real value
Output: a real value}
\usage{
tsal.q.from.shape(shape)
}
\description{
Calculate q parameter from shape parameter
Input: a real value
Output: a real value
}
