% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{tsal.qk.from.ss}
\alias{tsal.qk.from.ss}
\title{Calculate q & kappa parameters from shape & scale parameters
Input: two real values
Output: vector of two real values}
\usage{
tsal.qk.from.ss(shape, scale)
}
\description{
Calculate q & kappa parameters from shape & scale parameters
Input: two real values
Output: vector of two real values
}
