% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{tsal.scale.from.qk}
\alias{tsal.scale.from.qk}
\title{Calculate generalized-Pareto scale parameter from q and kappa parameters
Input: two real values
Output: a real value}
\usage{
tsal.scale.from.qk(q, kappa)
}
\description{
Calculate generalized-Pareto scale parameter from q and kappa parameters
Input: two real values
Output: a real value
}
