% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{tsal.shape.from.q}
\alias{tsal.shape.from.q}
\title{q-Exponential generator
Distribution-related (per R standards):
dtsal				Probability density
ptsal				Cumulative probability
qtsal				Quantiles
rtsal				Random variate generation
Parameter Conversion:
tsal.shape.from.q		Get shape parameter from q
tsal.scale.from.qk		Get scale parameter from q, kappa
tsal.q.from.shape		Get q from from shape parameter
tsal.kappa.from.ss		Get kappa from shape, scale
tsal.ss.from.qk		Get shape, scale from q, kappa (as pairs)
tsal.qk.from.ss		Get q, kappa from shape, scale (as pairs)
Users will generally have little reason to invoke these functions}
\usage{
tsal.shape.from.q(q)
}
\arguments{
\item{q}{a real value}
}
\value{
a real value
}
\description{
Calculate generalized-Pareto shape parameter from q parameter
}
