% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qExponentialGenerator.R
\name{tsal.ss.from.qk}
\alias{tsal.ss.from.qk}
\title{Calculate shape & scale parameters from q & kappa parameters
Input: two real values
Output: vector of two real values}
\usage{
tsal.ss.from.qk(q, kappa)
}
\description{
Calculate shape & scale parameters from q & kappa parameters
Input: two real values
Output: vector of two real values
}
