% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{var.teo.mle}
\alias{var.teo.mle}
\title{Theoretical variances}
\usage{
var.teo.mle(mu, theta)
}
\arguments{
\item{mu}{True value of mu}

\item{theta}{True value of theta}
}
\description{
Theoretical variances
}
