% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleQuantitiesR.R
\name{varTeoMle}
\alias{varTeoMle}
\title{Theoretical variances}
\usage{
varTeoMle(mu, theta)
}
\arguments{
\item{mu}{True value of mu}

\item{theta}{True value of theta}
}
\description{
Theoretical variances
}
