% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfmmQuantities.R
\name{mfmm.theo}
\alias{mfmm.theo}
\title{Theoretical quantities MFMM}
\usage{
mfmm.theo(mu, theta, u, v)
}
\arguments{
\item{mu}{true value of mu}

\item{theta}{true value of thete}

\item{u}{Tuning parameter}

\item{v}{Tuning parameter}
}
\description{
Theoretical quantities for the mfmm
}
